2_IntroToDistributedSystems/python-fast-api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
pydantic-settings
asyncpg
databases 
psycopg2
bcrypt